+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - [x] Open the .sprite file after it's created
 - [x] Command to update all sprites in solution
 - [x] Updated format to be more self-explanatory
-- [ ] Add logo watermark in .sprite files 
+- [x] Logo watermark in .sprite files 
 - [ ] Call Web Compiler to recompile when sprite updated
 
 Features that have a checkmark are complete and available for